csvBot/requirements.txt
streamlit>=1.32.0
pandas>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.2
langchain>=0.1.0
python-magic>=0.4.27
zipfile36>=0.1.3
